renaming from pychatbot to eddie
