+++ MANIFEST
@@ -1,9 +1,9 @@
 # file GENERATED by distutils, do NOT edit
 setup.cfg
 setup.py
-pychatbot/__init__.py
-pychatbot/bot.py
-pychatbot/endpoints/__init__.py
-pychatbot/endpoints/http.py
-pychatbot/endpoints/telegram.py
+eddie/__init__.py
+eddie/bot.py
+eddie/endpoints/__init__.py
+eddie/endpoints/http.py
+eddie/endpoints/telegram.py
 requirements.txt
