+++ eddie/__init__.py
