+++ eddie/bot.py
