+++ setup.cfg
@@ -3,7 +3,7 @@ current_version = 0.9.3
 commit = True
 tag = True
 
-[bumpversion:file:pychatbot/__init__.py]
+[bumpversion:file:eddie/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
